sentiment/requirements.txt
click==8.1.7
joblib==1.4.2
kafka-python-ng==2.2.3
nltk==3.9.1
regex==2024.11.6
setuptools==75.1.0
textblob==0.18.0.post0
tqdm==4.67.0
wheel==0.44.0
